requirements.txt
pyqt5
Pillow